refactor(dashboards): extract pure types + helpers from Network + Services monoliths
Partial close on #961.

Pulls the stateless surface out of the two big dashboard files into
sibling _lib/ modules — types, pure helpers, and the edge-kind
styling constants. The xyflow CustomNode / CustomEdge / NetworkLegend
React components stay in NetworkDashboard.tsx, and ServiceCard /
BundleCard stay in ServicesDashboard.tsx, because the issue
explicitly flags the canvas-state + render-cycle work as high risk
and we do not take that on without per-stage validation.

What lands:

dashboards/_lib/networkDashboard.ts (new, 165 LOC)
- GraphNodeData / LegacyPortMapping / HealthData types
- deriveNodeNameFromGraph, buildServiceEditHref
- DEFAULT / DOWN / DECLARED / OBSERVED edge color + dash tokens
- styleForEdgeKind, labelForEdgeKind

dashboards/_lib/servicesDashboard.ts (new, 130 LOC)
- MigrationPlan, LinkFormState, ApiLinkPayload, RawLinkPort, RawLinkVolume
- bundleWizardSteps, bundleSeverityClasses, MERGE_HELP_ID
- dedupeValidations

Sizes:
- NetworkDashboard.tsx: 2189 to 2083 LOC
- ServicesDashboard.tsx: 2029 to 1942 LOC

Next step toward the issue's full target (1500 LOC per file) is the
inner React component extraction — ServiceCard / BundleCard for
services, CustomNode / CustomEdge / NetworkLegend for network. Both
need a careful prop-surface design and per-stage live validation
against the dashboards; not in scope for this PR.
